Separate routers
Each algorithm has independent router
